# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HRM VNU-UET API Documentation",
    "version": "1.0.0",
    "description": "API specification for the Human Resource Management system of VNU-UET."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "V1 API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/reward/me": {
      "get": {
        "tags": [
          "Reward"
        ],
        "summary": "Get current user's rewards",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/reward/commendations": {
      "get": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "Get all commendations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "Create commendation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profile_id",
                  "award_level",
                  "award_name"
                ],
                "properties": {
                  "profile_id": {
                    "type": "integer"
                  },
                  "award_level": {
                    "type": "string",
                    "enum": [
                      "co_so",
                      "dhqg",
                      "bo",
                      "chinh_phu",
                      "nha_nuoc"
                    ]
                  },
                  "award_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/reward/commendations/{id}": {
      "put": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "Update commendation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "award_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "Delete commendation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/reward/commendations/{id}/attachments": {
      "post": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "Upload attachment for commendation",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Reward Commendations"
        ],
        "summary": "List attachments for commendation"
      }
    },
    "/reward/titles": {
      "get": {
        "tags": [
          "Reward Titles"
        ],
        "summary": "Get all titles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Reward Titles"
        ],
        "summary": "Create title",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profile_id",
                  "title_name",
                  "title_level",
                  "awarded_year"
                ],
                "properties": {
                  "profile_id": {
                    "type": "integer"
                  },
                  "title_name": {
                    "type": "string"
                  },
                  "title_level": {
                    "type": "string",
                    "enum": [
                      "unit",
                      "university",
                      "ministry"
                    ]
                  },
                  "awarded_year": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/reward/titles/{id}": {
      "put": {
        "tags": [
          "Reward Titles"
        ],
        "summary": "Update title",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reward Titles"
        ],
        "summary": "Delete title",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reward/discipline": {
      "get": {
        "tags": [
          "Reward Discipline"
        ],
        "summary": "Get all disciplinary records",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Reward Discipline"
        ],
        "summary": "Create disciplinary record",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profile_id",
                  "discipline_type",
                  "reason",
                  "decision_number",
                  "unit_name",
                  "issued_date"
                ],
                "properties": {
                  "profile_id": {
                    "type": "integer"
                  },
                  "discipline_type": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "decision_number": {
                    "type": "string"
                  },
                  "unit_name": {
                    "type": "string"
                  },
                  "issued_date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/reward/discipline/{id}": {
      "put": {
        "tags": [
          "Reward Discipline"
        ],
        "summary": "Update disciplinary record",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reward Discipline"
        ],
        "summary": "Delete disciplinary record",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}